Criacao de tela de criacao edicao de receita, passo de receita, ingredientes e helper de input de imagem

+++ Gourmetus/Gourmetus/Recipe/RecipeModel.swift
@@ -10,7 +10,7 @@ import Foundation
 struct RecipeModel {
     var name: String
     var desc: String?
-    var difficulty: Float
+    var difficulty: Int
     var imageData: Data?
     
     var steps: [StepModel]
